+++ packages/components/messages/index.js
@@ -1,4 +1,4 @@
-export { default as ErrorMessage } from '../../../src/components/messages/error-message';
-export { default as WarningMessage } from '../../../src/components/messages/warning-message';
+export { default as ErrorMessage } from './src/error-message';
+export { default as WarningMessage } from './src/warning-message';
 
 export { default as version } from './version';
