+++ packages/components/messages/src/error-message/error-message.spec.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render } from '../../../test-utils';
+import { render } from '../../../../../src/test-utils';
 import ErrorMessage from './error-message';
 
 describe('ErrorMessage', () => {
