+++ packages/components/messages/src/messages.visualroute.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { ErrorMessage, WarningMessage } from '@commercetools-frontend/ui-kit';
-import { Suite, Spec } from '../../../test/percy';
+import { Suite, Spec } from '../../../../test/percy';
 
 export const routePath = '/messages';
 // We group error and warning messages to save a snapshot.
