+++ packages/components/messages/src/warning-message/warning-message.spec.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render } from '../../../test-utils';
+import { render } from '../../../../../src/test-utils';
 import ErrorMessage from './warning-message';
 
 it('should render children', () => {
