+++ src/components/fields/localized-multiline-text-field/localized-multiline-text-field.story.js
@@ -9,8 +9,8 @@ import {
   select,
   object,
 } from '@storybook/addon-knobs/react';
+import { ErrorMessage } from '@commercetools-uikit/messages';
 import Section from '../../../../.storybook/decorators/section';
-import ErrorMessage from '../../messages/error-message';
 import Readme from './README.md';
 import * as icons from '../../icons';
 import LocalizedMultilineTextField from './localized-multiline-text-field';
