+++ src/components/inputs/creatable-select-input/creatable-select-input.form.story.js
@@ -5,9 +5,9 @@ import { action } from '@storybook/addon-actions';
 import { withKnobs, boolean } from '@storybook/addon-knobs/react';
 import Spacings from '@commercetools-uikit/spacings';
 import { PrimaryButton, SecondaryButton } from '@commercetools-uikit/buttons';
+import { ErrorMessage } from '@commercetools-uikit/messages';
 import Section from '../../../../.storybook/decorators/section';
 import FormikBox from '../../../../.storybook/decorators/formik-box';
-import ErrorMessage from '../../messages/error-message';
 import Readme from './README.md';
 import CreatableSelectInput from './creatable-select-input';
