+++ src/components/inputs/localized-money-input/localized-money-input.story.js
@@ -9,9 +9,8 @@ import {
   select,
   object,
 } from '@storybook/addon-knobs/react';
+import { ErrorMessage, WarningMessage } from '@commercetools-uikit/messages';
 import Section from '../../../../.storybook/decorators/section';
-import ErrorMessage from '../../messages/error-message';
-import WarningMessage from '../../messages/warning-message';
 import Readme from './README.md';
 import LocalizedMoneyInput from './localized-money-input';
