+++ src/components/inputs/localized-multiline-text-input/localized-multiline-text-input.story.js
@@ -9,9 +9,8 @@ import {
   select,
   object,
 } from '@storybook/addon-knobs/react';
+import { ErrorMessage, WarningMessage } from '@commercetools-uikit/messages';
 import Section from '../../../../.storybook/decorators/section';
-import ErrorMessage from '../../messages/error-message';
-import WarningMessage from '../../messages/warning-message';
 import Readme from './README.md';
 import LocalizedMultilineTextInput from './localized-multiline-text-input';
