+++ src/components/inputs/localized-rich-text-input/localized-rich-text-input.story.js
@@ -10,9 +10,8 @@ import {
   select,
   object,
 } from '@storybook/addon-knobs/react';
+import { ErrorMessage, WarningMessage } from '@commercetools-uikit/messages';
 import Section from '../../../../.storybook/decorators/section';
-import ErrorMessage from '../../messages/error-message';
-import WarningMessage from '../../messages/warning-message';
 import Readme from './README.md';
 import LocalizedRichTextInput from './localized-rich-text-input';
