+++ src/components/inputs/localized-text-input/localized-text-input.story.js
@@ -9,8 +9,8 @@ import {
   select,
   object,
 } from '@storybook/addon-knobs/react';
+import { ErrorMessage } from '@commercetools-uikit/messages';
 import Section from '../../../../.storybook/decorators/section';
-import ErrorMessage from '../../messages/error-message';
 import Readme from './README.md';
 import LocalizedTextInput from './localized-text-input';
