+++ src/components/inputs/money-input/money-input.form.story.js
@@ -7,7 +7,7 @@ import { withKnobs, number } from '@storybook/addon-knobs/react';
 import omitEmpty from 'omit-empty';
 import Spacings from '@commercetools-uikit/spacings';
 import { PrimaryButton, SecondaryButton } from '@commercetools-uikit/buttons';
-import ErrorMessage from '../../messages/error-message';
+import { ErrorMessage } from '@commercetools-uikit/messages';
 import Section from '../../../../.storybook/decorators/section';
 import FormikBox from '../../../../.storybook/decorators/formik-box';
 import Readme from './README.md';
