--- src/components/messages/error-message/README.md
@@ -1,23 +1,0 @@
-# ErrorMessage
-
-#### Description
-
-Represents an error message.
-
-## Usage
-
-```js
-import { ErrorMessage } from '@commercetools-frontend/ui-kit';
-
-<ErrorMessage>Something went wrong</ErrorMessage>;
-```
-
-#### Properties
-
-| Props      | Type   | Required | Values | Default | Description                             |
-| ---------- | ------ | :------: | ------ | ------- | --------------------------------------- |
-| `children` | `node` |    ✅    | -      | -       | Error message, either as string or node |
-
-Main functions and use cases are:
-
-- Display validation errors of input fields
