--- src/components/messages/warning-message/README.md
@@ -1,23 +1,0 @@
-# WarningMessage
-
-#### Description
-
-Represents an warning message.
-
-## Usage
-
-```js
-import { WarningMessage } from '@commercetools-frontend/ui-kit';
-
-<WarningMessage>This is a duplicate of {`${duplicate}`}</WarningMessage>;
-```
-
-#### Properties
-
-| Props      | Type   | Required | Values | Default | Description                               |
-| ---------- | ------ | :------: | ------ | ------- | ----------------------------------------- |
-| `children` | `node` |    ✅    | -      | -       | Warning message, either as string or node |
-
-Main functions and use cases are:
-
-- Display validation warnings of input fields
